Proposal for repository structure
Idea here is the keep the fixtures in root directory so that it's
independent on building / packaging tool.

See the example `java-libary` where the trick of including the fixtures
in .jar file is in usage of `resources` definition in `build` section to
point to custom path.

Do you have some prototype how it's used in some project? You mentioned
also some tests but I don't see them in repository.